requirements.txt
torch
torchvision
hydra-core
pytest
manim
einops
matplotlib
